# backend/docs/swagger.yaml
definitions:
  models.CommonResponse:
    properties:
      data: {}
      encrypted_data: {}
      error:
        type: boolean
      message:
        type: string
      meta: {}
      status:
        type: integer
    type: object
  models.MarriedCode:
    properties:
      created_at:
        $ref: '#/definitions/util.CustomTime'
      deleted_at:
        $ref: '#/definitions/util.CustomTime'
      id:
        type: integer
      married_code:
        type: string
      sam_id:
        type: string
      updated_at:
        $ref: '#/definitions/util.CustomTime'
    type: object
  models.RequestAuthLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.RequestCreateMarriedCode:
    properties:
      married_code:
        type: string
      sam_id:
        type: string
    type: object
  models.RequestCreateUser:
    properties:
      level:
        type: integer
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.RequestUpdateMarriedCode:
    properties:
      married_code:
        type: string
      sam_id:
        type: string
    type: object
  models.RequestUpdateUser:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  models.ResponseAuthLogin:
    properties:
      access_token:
        type: string
      expired_at:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  models.ResponseCreateUser:
    properties:
      id_user:
        type: integer
      level:
        type: integer
      name:
        type: string
      username:
        type: string
    type: object
  util.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: Register a login in the system
      parameters:
      - description: Login creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestAuthLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ResponseAuthLogin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      summary: Create login
      tags:
      - Login
  /v1/married-codes:
    get:
      consumes:
      - application/json
      description: Retrieve all marriedcodes with pagination
      parameters:
      - description: Limit the number of marriedcodes per page (min 1 and max 100)
        example: 10
        in: query
        name: limit
        required: true
        type: integer
      - description: Page number for pagination (min 1)
        example: 1
        in: query
        name: page
        required: true
        type: integer
      - description: Search term for sam_id and married_code
        in: query
        name: search
        type: string
      - description: Order for sam_id and married_code
        in: query
        name: order_by
        type: string
      - description: Sort for asc and desc
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MarriedCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get All MarriedCodes
      tags:
      - MarriedCode
    post:
      consumes:
      - application/json
      description: Create marriedcode in the system
      parameters:
      - description: MarriedCode creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestCreateMarriedCode'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MarriedCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Create MarriedCode
      tags:
      - MarriedCode
  /v1/married-codes/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a married code by ID
      parameters:
      - description: Married Code ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully soft deleted married code
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "404":
          description: Married code not found
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Soft Delete a Married Code
      tags:
      - MarriedCode
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific married code by ID
      parameters:
      - description: Married Code ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with married code details
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "404":
          description: Married code not found
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Get Married Code Details
      tags:
      - MarriedCode
    put:
      consumes:
      - application/json
      description: Update an existing married code by ID
      parameters:
      - description: Married Code ID
        in: path
        name: id
        required: true
        type: integer
      - description: Married code update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestUpdateMarriedCode'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated married code
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "400":
          description: Invalid request body or duplicate married_code
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized - Invalid or missing token
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "404":
          description: Married code not found
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update a Married Code
      tags:
      - MarriedCode
  /v1/user:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestCreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ResponseCreateUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      summary: Create a new user
      tags:
      - User
  /v1/user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete a user from the system
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update data user in the system
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: User update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RequestUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - User
  /v1/user/search:
    get:
      consumes:
      - application/json
      description: Search in the system
      parameters:
      - description: query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseCreateUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.CommonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.CommonResponse'
      security:
      - BearerAuth: []
      summary: Search
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"